Grouped exporting to separate class
