+++ lib/hammer_cli_foreman_tasks.rb
@@ -2,6 +2,7 @@ module HammerCLIForemanTasks
   require 'hammer_cli_foreman_tasks/helper'
   require 'hammer_cli_foreman_tasks/task_progress'
   require 'hammer_cli_foreman_tasks/async_command'
+  require 'hammer_cli_foreman_tasks/exporter'
   require 'hammer_cli_foreman_tasks/viewer'
   require 'hammer_cli_foreman_tasks/dynflow_binding'
   require 'hammer_cli_foreman_tasks/action'
