+++ lib/hammer_cli_foreman_tasks/helper.rb
@@ -24,10 +24,5 @@ module HammerCLIForemanTasks
       end
     end
 
-    def compress(target, dest='.')
-      archive = "#{dest}/#{target}.tar.gz"
-      tgz = Zlib::GzipWriter.new(File.open(archive, 'wb'))
-      Archive::Tar::Minitar.pack(target, tgz)
-    end
   end
 end
